# arthurclive_swagger.yaml
swagger: '2.0'
info:
  description: Arthur Clive Web Api
  version: 1.0.0
  title: Arthur Clive Swagger
  contact:
    email: raguvarthan.n@turingminds.com
host: arhturclive.swagger.io
basePath: /v1
...
schemes:
  - http
paths:
  /api/product:
    get:
      tags:
        - ProductController
      operationId: getProducts
      produces:
        - application/xml
        - application/json
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ProductDetails'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
    post:
      tags:
        - ProductController
      operationId: insertProduct
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: body
          description: Object that need to be added to the DB
          required: true
          schema:
            $ref: '#/definitions/Product'
      responses:
        '200':
          description: Inserted
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
  '/api/product/{productSKU}':
    delete:
      tags:
        - ProductController
      operationId: deleteProduct
      produces:
        - application/xml
        - application/json
      parameters:
        - name: productSKU
          in: path
          description: SKU of the product that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: Deleted
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
        '404':
          description: Product Not Found
          schema:
            $ref: '#/definitions/ResponseData'
  /api/category:
    get:
      tags:
        - CategoryController
      operationId: getCategories
      produces:
        - application/xml
        - application/json
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/CategoryDetails'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
    post:
      tags:
        - CategoryController
      operationId: insertCategory
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: product
          description: Object that need to be added to the DB
          required: true
          schema:
            $ref: '#/definitions/Category'
      responses:
        '200':
          description: Inserted
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
  '/api/category/{productFor}/{productType}':
    delete:
      tags:
        - CategoryController
      operationId: deleteCategory
      produces:
        - application/xml
        - application/json
      parameters:
        - name: productFor
          in: path
          description: productFor of the category that needs to be deleted
          required: true
          type: string
        - name: productType
          in: path
          description: productType of the category that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: Deleted
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
        '404':
          description: Product Not Found
          schema:
            $ref: '#/definitions/ResponseData'
...
definitions:
  ResponseData:
    type: object
    properties:
      Code:
        type: number
      Message:
        type: string
      Data:
        type: object
      Content:
        type: object
  Product:
    type: object
    properties:
      ProductFor:
        type: string
      ProductType:
        type: string
      ProductDesign:
        type: string
      ProductBrand:
        type: string
      ProductPrice:
        type: number
      ProductDiscount:
        type: number
      ProductStock:
        type: number
      ProductSize:
        type: string
      ProductMaterial:
        type: string
      ProductRating:
        type: number
      ProductReviews:
        $ref: '#/definitions/Reviews'
      ProductColour:
        type: string
      RefundApplicable:
        type: boolean
      ReplacementApplicable:
        type: boolean
      ProductDescription:
        type: string
  ProductDetails:
    type: object
    properties:
      Id:
        type: number
      ProductSKU:
        type: string
      MinioObject_URl:
        type: string
      ProductFor:
        type: string
      ProductType:
        type: string
      ProductDesign:
        type: string
      ProductBrand:
        type: string
      ProductPrice:
        type: number
      ProductDiscount:
        type: number
      ProductDiscountPrice:
        type: string
      ProductStock:
        type: number
      ProductSize:
        type: string
      ProductMaterial:
        type: string
      ProductRating:
        type: number
      ProductReviews:
        $ref: '#/definitions/Reviews'
      ProductColour:
        type: string
      RefundApplicable:
        type: boolean
      ReplacementApplicable:
        type: boolean
      ProductDescription:
        type: string
  Reviews:
    type: object
    properties:
      Name:
        type: string
      Comment:
        type: string
  Category:
    type: object
    properties:
      ProductFor:
        type: string
      ProductType:
        type: string
      Description:
        type: string
  CategoryDetails:
    type: object
    properties:
      Id:
        type: number
      ProductFor:
        type: string
      ProductType:
        type: string
      MinioObject_URL:
        type: string
      Description:
        type: string
...